State recoverable from persisted commit log
Using function recoverState(), we parse the commit log file and for each line, run the appropriate command and update state.
As of now, it's only usable when a machine has no state (i.e. when we first start up our servers.) Code still needs to be improved
to update SR's commit logs so that recoverState only runs commands that are more recent than what's already in commit logs (right
now, SR's have state updates, but commit log is not updated)